Updates Databse File
You will need to add the foreign key manually in adminer unless you can figure out what's wrong. Then again you may not have an issue. If you are able to add everything in the SQL command go for it. SQL is funky sometimes. The good thing is we have a work around.